PORTLETBRIDGE-22: Update PortletBridge to latest draft (revision 11) PORTLETBRIDGE-21: Wrong leading slash in BRIDGE_SERVICE_CLASSPATH /GenericFacesPortlet
These issues have been fixes under a patch submitted to 
PORTLETBRIDGE-22 by Michael Freedman.  I had to remove the 
BridgeRenderFilter for reasons documented in the bug and have made an
inquiry to the EG on how this needs to be resolved from a spec 
standpoint.  Likely it'll be left up to an implementation detail for
the Apache community at which point I'll generate another bug.

Thanks Michael for all your work on this patch.

git-svn-id: https://svn.apache.org/repos/asf/myfaces/portlet-bridge/trunk@610847 13f79535-47bb-0310-9956-ffa450edef68